Simplification - RootModule is Baubit.DI.ARootModule. Baubit.Autofac.DI.IModule hides Load(IServiceCollection)
