+++ Baubit.Autofac.Test/DI/Setup/Configuration.cs
@@ -1,6 +1,6 @@
 ﻿namespace Baubit.Autofac.Test.DI.Setup
 {
-    public class Configuration : Baubit.Autofac.DI.AConfiguration
+    public class Configuration : Baubit.DI.AConfiguration
     {
         public string SomeString { get; init; }
         public string SomeSecretString { get; init; }
