+++ Baubit.Autofac/Baubit.Autofac.csproj
@@ -26,7 +26,7 @@
   <ItemGroup>
     <PackageReference Include="Autofac" Version="8.3.0" />
     <PackageReference Include="Autofac.Extensions.DependencyInjection" Version="10.0.0" />
-    <PackageReference Include="Baubit" Version="2025.22.3" />
+    <PackageReference Include="Baubit" Version="2025.23.1" />
   </ItemGroup>
 
 </Project>
