--- Baubit.Autofac/DI/AConfiguration.cs
@@ -1,6 +1,0 @@
-﻿namespace Baubit.Autofac.DI
-{
-    public abstract class AConfiguration : Baubit.DI.AConfiguration
-    {
-    }
-}
