+++ Baubit.Autofac/DI/RootModule.cs
@@ -1,5 +1,4 @@
 ﻿using Autofac;
-using Autofac.Builder;
 using Autofac.Extensions.DependencyInjection;
 using Baubit.Configuration;
 using Baubit.DI;
@@ -7,48 +6,9 @@ using Baubit.Traceability;
 using FluentResults;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
 
 namespace Baubit.Autofac.DI
 {
-
-    public abstract class ARootModuleConfiguration : AConfiguration
-    {
-
-    }
-    public sealed class RootModuleConfiguration : ARootModuleConfiguration
-    {
-        public ContainerBuildOptions ContainerBuildOptions { get; init; }
-    }
-
-    public abstract class ARootModule<TConfiguration,
-                                      TServiceProviderFactory,
-                                      TContainerBuilder> : AModule<TConfiguration>, IRootModule where TConfiguration : ARootModuleConfiguration
-                                                                                                  where TServiceProviderFactory : IServiceProviderFactory<TContainerBuilder>
-                                                                                                  where TContainerBuilder : notnull
-    {
-        protected ARootModule(ConfigurationSource configurationSource) : base(configurationSource)
-        {
-        }
-
-        protected ARootModule(IConfiguration configuration) : base(configuration)
-        {
-        }
-
-        protected ARootModule(TConfiguration configuration, List<AModule> nestedModules, List<IConstraint> constraints) : base(configuration, nestedModules, constraints)
-        {
-        }
-
-        public Result<THostApplicationBuilder> UseConfiguredServiceProviderFactory<THostApplicationBuilder>(THostApplicationBuilder hostApplicationBuilder) where THostApplicationBuilder : IHostApplicationBuilder
-        {
-            hostApplicationBuilder.ConfigureContainer(GetServiceProviderFactory(), GetConfigureAction());
-            return hostApplicationBuilder;
-        }
-
-        protected abstract TServiceProviderFactory GetServiceProviderFactory();
-        protected abstract Action<TContainerBuilder> GetConfigureAction();
-
-    }
     public sealed class RootModule : ARootModule<RootModuleConfiguration, AutofacServiceProviderFactory, ContainerBuilder>, IModule
     {
         public RootModule(ConfigurationSource configurationSource) : base(configurationSource)
@@ -70,7 +30,7 @@ namespace Baubit.Autofac.DI
                              .ThrowIfFailed();
         }
 
-        public override void Load(ContainerBuilder containerBuilder)
+        public void Load(ContainerBuilder containerBuilder)
         {
             var baubitModules = new List<Baubit.DI.IModule>();
             var baubitAutofacModules = new List<IModule>();
